# OpenAPI.yml
openapi:  3.0.3
info:
  title: CAPSTONE PROJECT kamarRS
  version: 1.0.0
  description:  Dokumentasi Project 
  contact:
    name: kamarRS team
    email: lzfk0x9a@duck.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/kamar-rs/kamar-rs/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Anti-Gen/CAPSTONE-PORJECT-kamarRS/1.0.0
  - description: kamarRS server
    url: http://35.198.253.37:80
  - description: kamarRS domain
    url: https://rawatinap.online
...
paths:
...
  /registrations:
    post:
      tags:
        - Beds Registrations
      summary: add patient registration
      description: Menambahkan data registrasi pasien icu/rawat inap
      security: 
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                hospital_id: 3
                patient_id: 255
      responses:
        '201':
          description: penambahan data berhasil
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success create registrations"
                  data: 
                    hospital_id: 3
                    patient_id: 255
                    biaya_registrasi: 25000
                    kode_daftar: "order-912"
                    status_pembayaran: "belum dibayar"
                    
        '500':
          description: penambahan data gagal
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "failed"
                  message: "failed create registrations"
  /hospitals/{hospitalId}/registrations:
    get:
      tags:
        - Beds Registrations
      summary: Get beds registrations by hospital id
      description: Menampilkan data pendaftaran berdasarkan id rumah sakit
      security:
        - JWTAuth: []
      parameters:
        - name: hospital id
          description: "filter beds registrations by hospital_id"
          required: true
          in: path
          schema: 
            type: integer
            example: 1
        - name: page
          description: "shows result per page"
          required: false
          in: query
          schema: 
            type: integer
            example: 1
      responses: 
        '201':
          description: List of beds registrations by hospital id
          content: 
            application/json:
             schema:
              type: object
              example:
                status: "success"
                message: "success get all beds registrations"
                data: 
                  - id: 1
                    hospital_id: 3
                    patient_id: 176
                    bed_id: 4
                    status_pasien: "pemeriksaan awal"
                    biaya_registrasi: 25000
                    kode_daftar: "Postman-1578568851"
                    link-pembayaran: "masih tidak tahu linknya bagaimana"
                    status-pembayaran: "success"
                  - id: 5
                    hospital_id: 3
                    patient_id: 233
                    bed_id: 10
                    status_pasien: "rawat inap"
                    biaya_registrasi: 25000
                    kode_daftar: "Postman-1578568851"
                    link-pembayaran: "masih tidak tahu linknya bagaimana"
                    status-pembayaran: "success"
                  - id: 15
                    hospital_id: 3
                    patient_id: 287
                    bed_id: 17
                    status_pasien: "dikembalikan ke wali"
                    biaya_registrasi: 25000
                    kode_daftar: "Postman-1578568851"
                    link-pembayaran: "masih tidak tahu linknya bagaimana"
                    status-pembayaran: "success"
        '500':
          $ref: '#/components/responses/500'
...
  /payments/{kode_daftar}:
    put: 
      tags:
        - Beds Registrations
      summary: payment for bed registration
      description: Menambahkan data registrasi pasien icu/rawat inap
      security: 
        - JWTAuth: []
      parameters:
        - name: kode_daftar
          description: "kode daftar"
          required: true
          in: path
          schema:
            description: "kode daftar"
            type: string
            example: "order-134"
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                metode_pembayaran:
                  type: string
                  enum:
                    - "transer_va_bca"
                    - "transer_va_bri"
                    - "transer_va_bni"
                    - "transer_va_permata"
                    - "qris"
              example: 
                metode_pembayaran: "transfer_va_bca"
      responses:
        '200':
          description: "success update payment for bed reservation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithoutData"
              example:
                status: "success"
                message: "success create payment"
                data: 
                  status_pasien: "waiting list"
                  biaya_registrasi: 25000
                  Kode_daftar: "order-Hssy4"
                  payment_method: "transfer_va_bca"
                  link_pembayaran: "https://simulator.sandbox.midtrans.com/bca/va/index"
                  transaction_id: "6ffa1be9-28ae-4f44-ba88-e71f75d5899d"
                  virtual_account: "52185383264"
                  bank_penerima: "bank bca"
                  waktu_kedaluarsa: "2022-12-31 17:15:49 +0000 UTC"
                  qr_string: "lkfjaogfnbalksfio9wa"
                  status_pembayaran: "pending"
                  hospital_id: 1
        '400':
          $ref: "#/components/responses/400"
        "401" :
          $ref : "#/components/responses/401"  
        "404" :
          $ref : "#/components/responses/404"  
        '500':
          $ref: "#/components/responses/500"
    get:
      tags:
        - Beds Registrations
      summary: get payment information for bed registration
      description: Menambahkan data registrasi pasien icu/rawat inap
      security: 
        - JWTAuth: []
      parameters:
        - name: kode_daftar
          description: "kode daftar"
          required: true
          in: path
          schema:
            description: "kode daftar"
            type: string
            example: "order-134"
      responses:
        "200":
          description: "success get payment info for order id"
          content:
            application/json:
              schema:
                type: object
                properties:
                  kode_daftar:
                    type: string 
                  metode_pembayaran:
                    type: string
                  biaya_registrasi:
                    type: integer
                    format: int64
                  link_pembayaran:
                    type: string
                  qr_string:
                    type: string
                  status_pembayaran:
                    type: string
              example:
                kode_daftar: "order-134"
                metode_pembayaran: "qris"
                biaya_registrasi: 25000
                link_pembayaran: "https:midtras/bayar"
                qr_string: "string untuk generate qr code"
                status_pembayaran: "belum dibayar"

##############################################################
##############################################################
  /midtrans:
    put:
      tags:
        - Invoice
      summary: callback
      description: callback pembayaran midtrans
      security:
        - JWTAuth: []
      responses:
        "200":
          description: Success update status.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ResponseObject"
              example:
                message: "success update status"
                data :
                  id_booking: 1
                  kode_daftar: "order-101"
                  gross_amount: 60000
                  status: "complete"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "user Token"
  schemas:
...
    ResponseWithoutData:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    ResponseObject:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data: 
          type: object
...
  responses:
    '400':
      description: "Wrong Request From User"
      content:
        "application/json":
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
            properties:
              message:
                type: string
          example:
            message: "Process Failed. Please check your input."
    '401':
      description: "Authorization information is missing or invalid."
      content:
        "application/json":
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
            properties:
              message:
                type: string
          example:
            message: "Process Failed. User not authorized for this request."
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "data not found"
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "internal server error"